# `vector_algorithms.cpp`: Make AVX2 path avoid SSE path and fallback