@StephanTLavavej 2024-04-11
Thanks, this looks awesome! I pushed an essentially stylistic change to how we deal with `_Traits::_Shift` - I think the result is clearer, but please meow if you disagree. :heart_eyes_cat: